apps/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.39
asyncpg==0.30.0
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
redis==5.2.1
elasticsearch[async]==8.17.2
httpx==0.28.1
email-validator==2.2.0
python-slugify==8.0.4
tenacity==9.0.0
orjson==3.10.16
